+++ src/bim-links/bentley-ram/IdeaStatiCa.RamToIdeaApp/App.xaml.cs
@@ -1,8 +1,7 @@
 ﻿using IdeaStatiCa.Plugin;
 using IdeaStatiCa.PluginLogger;
 using IdeaStatiCa.RamToIdeaApp.Services;
-using IdeaStatiCa.RamToIdeaApp.View;
-using IdeaStatiCa.RamToIdeaApp.ViewModels;
+
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using System;
@@ -19,11 +18,7 @@ namespace IdeaStatiCa.RamToIdeaApp
 
 		public App()
 		{
-			IConfiguration configuration = BuildConfiguration();
 			var services = new ServiceCollection();
-
-			services.AddSingleton<IConfiguration>(configuration);
-
 			services.AddSingleton<IPluginLogger>(serviceProvider =>
 			{
 				SerilogFacade.Initialize();
@@ -31,37 +26,27 @@ namespace IdeaStatiCa.RamToIdeaApp
 			});
 
 			services.AddTransient<IProjectService, ProjectService>();
-
-			services.AddTransient<MainWindow>(serviceProvider => new MainWindow
-			{
-				DataContext = serviceProvider.GetRequiredService<MainWindowViewModel>()
-			});
-
-			services.AddTransient<MainWindowViewModel>();
-
-
+			services.AddTransient<IStartupService, StartupService>();
 
 			serviceProvider = services.BuildServiceProvider();
 		}
 
-		protected override void OnStartup(StartupEventArgs e)
+		protected override async void OnStartup(StartupEventArgs e)
 		{
-			var mainWindow = serviceProvider.GetRequiredService<MainWindow>();
-			mainWindow.Show();
 			base.OnStartup(e);
+
+			var startupService = serviceProvider.GetRequiredService<IStartupService>();
+
+			// Call the async method and wait for it
+			await startupService.RunCheckbotAsync();
+
+			// Shut down the application after completion
+			Shutdown();
 		}
 
 		protected override void OnExit(ExitEventArgs e)
 		{
 			base.OnExit(e);
-		}
-
-		public static IConfigurationRoot BuildConfiguration()
-		{
-			return new ConfigurationBuilder()
-				.AddJsonFile("appsettings.json")
-				.AddEnvironmentVariables()
-				.Build();
 		}
 	}
