+++ src/bim-links/bentley-ram/IdeaStatiCa.RamToIdeaApp/IdeaStatiCa.RamToIdeaApp.csproj
@@ -5,7 +5,11 @@
 		<TargetFramework>net6.0-windows</TargetFramework>
 		<Configurations>Release_IdeaStatiCa_Internal;Debug_IdeaStatiCa_Internal;Debug;Release</Configurations>
 		<UseWPF>true</UseWPF>
+		<ApplicationIcon>Resources\RAM icon.ico</ApplicationIcon>
 	</PropertyGroup>
+	<ItemGroup>
+	  <None Remove="Resources\RAM icon.ico" />
+	</ItemGroup>
 
 	<ItemGroup>
 		<PackageReference Include="CommunityToolkit.Mvvm" />
@@ -21,9 +25,9 @@
 	</ItemGroup>
 
 	<ItemGroup>
-		<None Update="appsettings.json">
-			<CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-		</None>
+	  <Resource Include="Resources\RAM icon.ico">
+	    <CopyToOutputDirectory>Never</CopyToOutputDirectory>
+	  </Resource>
 	</ItemGroup>
 
 </Project>
