+++ src/bim-links/bentley-ram/IdeaStatiCa.RamToIdeaApp/Services/IStartupService.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace IdeaStatiCa.RamToIdeaApp.Services
+{
+	public interface IStartupService
+	{
+		Task RunCheckbotAsync();
+	}
+}
