--- src/bim-links/bentley-ram/IdeaStatiCa.RamToIdeaApp/ViewModels/ViewModelBase.cs
@@ -1,8 +1,0 @@
-﻿using CommunityToolkit.Mvvm.ComponentModel;
-
-namespace IdeaStatiCa.RamToIdeaApp.ViewModels
-{
-	public class ViewModelBase : ObservableObject
-	{
-	}
-}
