--- src/bim-links/bentley-ram/IdeaStatiCa.RamToIdeaApp/Views/MainWindow.xaml
@@ -1,25 +1,0 @@
-﻿<Window x:Class="IdeaStatiCa.RamToIdeaApp.View.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:IdeaStatiCa.RamToIdeaApp"
-        mc:Ignorable="d"
-        Title="RAM To Idea" Height="450" Width="800">
-	<DockPanel LastChildFill="True">
-		<ToolBarTray DockPanel.Dock="Top">
-			<ToolBar Band="1" DockPanel.Dock="Top">
-				<Button Command="{Binding RunCheckBotCommand, Mode=OneWay}">Run Checkbot</Button>
-			</ToolBar>
-		</ToolBarTray>
-		<StackPanel DockPanel.Dock="Bottom">
-			<TextBlock Text="{Binding Status}"></TextBlock>
-		</StackPanel>
-		<Grid>
-			<Grid.ColumnDefinitions>
-				<ColumnDefinition></ColumnDefinition>
-				<ColumnDefinition></ColumnDefinition>
-			</Grid.ColumnDefinitions>
-		</Grid>
-	</DockPanel>
-</Window>
