--- src/bim-links/bentley-ram/IdeaStatiCa.RamToIdeaApp/Views/MainWindow.xaml.cs
@@ -1,15 +1,0 @@
-﻿using System.Windows;
-
-namespace IdeaStatiCa.RamToIdeaApp.View
-{
-	/// <summary>
-	/// Interaction logic for MainWindow.xaml
-	/// </summary>
-	public partial class MainWindow : Window
-	{
-		public MainWindow()
-		{
-			InitializeComponent();
-		}
-	}
-}
