--- src/bim-links/bentley-ram/IdeaStatiCa.RamToIdeaApp/appsettings.json
@@ -1,3 +1,0 @@
-{
-
-}
